third_party/WebKit/LayoutTests/dom/document_type/remove-expected.txt

This tests the DOM 4 remove method on a DocumentType node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.childNodes.length is 2
PASS document.childNodes.length is 1
PASS document.childNodes.length is 1
PASS successfullyParsed is true

TEST COMPLETE
